--DRAFF --new-api-order

+++ src/main/java/com/fanstatic/dto/model/product/ProductDTO.java
@@ -18,7 +18,7 @@ public class ProductDTO extends ResponseDataDTO {
 
     private String name;
 
-    private BigInteger price;
+    private Long price;
 
     private byte active;
